+++ pkg/scalers/liiklus_scaler.go
@@ -46,6 +46,17 @@ const (
 	liiklusMetricType                       = "External"
 )
 
+var (
+	// ErrLiiklusNoTopic is returned when "topic" in the config is empty.
+	ErrLiiklusNoTopic = errors.New("no topic provided")
+
+	// ErrLiiklusNoAddress is returned when "address" in the config is empty.
+	ErrLiiklusNoAddress = errors.New("no liiklus API address provided")
+
+	// ErrLiiklusNoGroup is returned when "group" in the config is empty.
+	ErrLiiklusNoGroup = errors.New("no consumer group provided")
+)
+
 // NewLiiklusScaler creates a new liiklusScaler scaler
 func NewLiiklusScaler(config *ScalerConfig) (Scaler, error) {
 	metricType, err := GetMetricTargetType(config)
@@ -157,7 +168,7 @@ func parseLiiklusMetadata(config *ScalerConfig) (*liiklusMetadata, error) {
 	if val, ok := config.TriggerMetadata[liiklusActivationLagThresholdMetricName]; ok {
 		t, err := strconv.ParseInt(val, 10, 64)
 		if err != nil {
-			return nil, fmt.Errorf("error parsing %s: %s", liiklusActivationLagThresholdMetricName, err)
+			return nil, fmt.Errorf("error parsing %s: %w", liiklusActivationLagThresholdMetricName, err)
 		}
 		activationLagThreshold = t
 	}
@@ -173,11 +184,11 @@ func parseLiiklusMetadata(config *ScalerConfig) (*liiklusMetadata, error) {
 
 	switch {
 	case config.TriggerMetadata["topic"] == "":
-		return nil, errors.New("no topic provided")
+		return nil, ErrLiiklusNoTopic
 	case config.TriggerMetadata["address"] == "":
-		return nil, errors.New("no liiklus API address provided")
+		return nil, ErrLiiklusNoAddress
 	case config.TriggerMetadata["group"] == "":
-		return nil, errors.New("no consumer group provided")
+		return nil, ErrLiiklusNoGroup
 	}
 
 	return &liiklusMetadata{