+++ CHANGELOG.md
@@ -73,6 +73,7 @@ New deprecation(s):
 ### Other
 
 - **RabbitMQ Scaler:** Move from `streadway/amqp` to `rabbitmq/amqp091-go` ([#4004](https://github.com/kedacore/keda/pull/4039))
+- **General:** Compare error with `errors.Is` ([#4004](https://github.com/kedacore/keda/pull/4004))
 
 ## v2.9.1